feat(workflow): migrate ai, answer, information-extractor and text-classifier to the catalog
