+++ apps/web/src/components/workflow/nodes/core/answer/types.ts
@@ -1,47 +1,19 @@
 // apps/web/src/components/workflow/nodes/core/answer/types.ts
 
-import type { BaseNodeData, ExecutionResult, SpecificNode } from '../../../types'
+import type { CatalogAnswerNodeData } from '@auxx/lib/workflow-engine/client'
+import type { ExecutionResult, SpecificNode } from '../../../types'
+import type { NodeType } from '../../../types/node-types'
+
+// The data half moved to the node catalog (node-catalog Phase 1 —
+// `@auxx/lib/workflow-engine/catalog/nodes/answer`); AnswerNodeData narrows
+// `type` to the web NodeType enum, same as BaseNodeData does over its lib
+// counterpart.
 
 /**
  * Node data for answer nodes with flattened structure
  */
-export interface AnswerNodeData extends BaseNodeData {
-  // 'new' = compose fresh email, 'reply' = reply to sender, 'replyAll' = reply to all
-  messageType: 'new' | 'reply' | 'replyAll'
-
-  integrationId?: string // Required for 'new', auto-derived for reply/replyAll
-  recordId?: string // Format: "entityDefinitionId:id" (e.g. "thread:abc123", "message:xyz789")
-  // Required for reply/replyAll. Replaces old resourceId + resourceType.
-
-  // Per-field auto-resolve toggles (reply/replyAll only, ignored for 'new')
-  // true (default): system auto-resolves from thread at execution time
-  // false: user provides the value explicitly
-  toIsAuto?: boolean
-  ccIsAuto?: boolean
-  bccIsAuto?: boolean
-  subjectIsAuto?: boolean
-
-  to?: string[]
-  toModes?: boolean[]
-  cc?: string[]
-  ccModes?: boolean[]
-  bcc?: string[]
-  bccModes?: boolean[]
-  text: string
-  subject?: string
-  /**
-   * Attachment picker rows — a constant row is a `file:<id>` reference, a
-   * variable row a `{{…}}` reference. `attachmentFilesModes` marks which is
-   * which, positionally. The engine resolves both to file ids and hands them to
-   * the send path as `attachmentIds`.
-   */
-  attachmentFiles?: string[]
-  attachmentFilesModes?: boolean[]
-  saveAsDraft?: boolean // When true, creates a draft instead of sending
-  // Per-node send behavior override: default = dry-run during test runs,
-  // live = always send, dry_run = never send, draft = save as thread draft
-  test_behavior?: 'default' | 'live' | 'dry_run' | 'draft'
-  fieldModes?: Record<string, boolean> // Track constant/variable mode per field
+export interface AnswerNodeData extends CatalogAnswerNodeData {
+  type: NodeType
 }
 
 /**
